Safari-611.1.10/LayoutTests/fast/text/emoji-num-glyphs.html

<!DOCTYPE html>
<html>
<head>
<style>
#console {
	font: 16px Times;
}
</style>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body style="font: 100px Times;">
<span id="control">&#x26F9;</span>
<span class="testElement">&#x1F468;&zwj;&#x1F466;</span>
<span class="testElement">&#x1F468;&zwj;&#x1F466;&zwj;&#x1F466;</span>
<span class="testElement">&#x1F468;&zwj;&#x1F467;</span>
<span class="testElement">&#x1F468;&zwj;&#x1F467;&zwj;&#x1F466;</span>
<span class="testElement">&#x1F468;&zwj;&#x1F467;&zwj;&#x1F467;</span>
<span class="testElement">&#x1F469;&zwj;&#x1F466;</span>
<span class="testElement">&#x1F469;&zwj;&#x1F466;&zwj;&#x1F466;</span>
<span class="testElement">&#x1F469;&zwj;&#x1F467;</span>
<span class="testElement">&#x1F469;&zwj;&#x1F467;&zwj;&#x1F466;</span>
<span class="testElement">&#x1F469;&zwj;&#x1F467;&zwj;&#x1F467;</span>



<span class="testElement">&#x1F3F3;&#xFE0F;&zwj;&#x1F308;</span>
<span class="testElement">&#x1F3F3;&zwj;&#x1F308;</span>



<span class="testElement">&#x1F3C3;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3C3;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3C3;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3C3;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3C3;&#x1F3FB;</span>
<span class="testElement">&#x1F3C3;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3C3;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3C3;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3C3;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3C3;&#x1F3FC;</span>
<span class="testElement">&#x1F3C3;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3C3;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3C3;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3C3;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3C3;&#x1F3FD;</span>
<span class="testElement">&#x1F3C3;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3C3;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3C3;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3C3;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3C3;&#x1F3FE;</span>
<span class="testElement">&#x1F3C3;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3C3;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3C3;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3C3;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3C3;&#x1F3FF;</span>
<span class="testElement">&#x1F3C3;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3C3;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3C3;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3C3;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3C3;</span>
<span class="testElement">&#x1F3C4;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3C4;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3C4;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3C4;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3C4;&#x1F3FB;</span>
<span class="testElement">&#x1F3C4;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3C4;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3C4;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3C4;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3C4;&#x1F3FC;</span>
<span class="testElement">&#x1F3C4;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3C4;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3C4;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3C4;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3C4;&#x1F3FD;</span>
<span class="testElement">&#x1F3C4;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3C4;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3C4;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3C4;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3C4;&#x1F3FE;</span>
<span class="testElement">&#x1F3C4;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3C4;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3C4;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3C4;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3C4;&#x1F3FF;</span>
<span class="testElement">&#x1F3C4;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3C4;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3C4;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3C4;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3C4;</span>
<span class="testElement">&#x1F3CA;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3CA;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3CA;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3CA;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3CA;&#x1F3FB;</span>
<span class="testElement">&#x1F3CA;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3CA;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3CA;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3CA;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3CA;&#x1F3FC;</span>
<span class="testElement">&#x1F3CA;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3CA;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3CA;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3CA;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3CA;&#x1F3FD;</span>
<span class="testElement">&#x1F3CA;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3CA;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3CA;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3CA;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3CA;&#x1F3FE;</span>
<span class="testElement">&#x1F3CA;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3CA;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3CA;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3CA;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3CA;&#x1F3FF;</span>
<span class="testElement">&#x1F3CA;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3CA;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3CA;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3CA;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3CA;</span>
<span class="testElement">&#x1F3CB;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3CB;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3CB;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3CB;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3CB;&#x1F3FB;</span>
<span class="testElement">&#x1F3CB;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3CB;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3CB;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3CB;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3CB;&#x1F3FC;</span>
<span class="testElement">&#x1F3CB;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3CB;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3CB;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3CB;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3CB;&#x1F3FD;</span>
<span class="testElement">&#x1F3CB;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3CB;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3CB;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3CB;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3CB;&#x1F3FE;</span>
<span class="testElement">&#x1F3CB;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3CB;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3CB;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3CB;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3CB;&#x1F3FF;</span>
<span class="testElement">&#x1F3CB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3CB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3CB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3CB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3CB;</span>
<span class="testElement">&#x1F3CC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3CC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3CC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3CC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3CC;</span>
<span class="testElement">&#x1F46E;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F46E;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F46E;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F46E;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F46E;&#x1F3FB;</span>
<span class="testElement">&#x1F46E;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F46E;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F46E;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F46E;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F46E;&#x1F3FC;</span>
<span class="testElement">&#x1F46E;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F46E;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F46E;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F46E;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F46E;&#x1F3FD;</span>
<span class="testElement">&#x1F46E;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F46E;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F46E;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F46E;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F46E;&#x1F3FE;</span>
<span class="testElement">&#x1F46E;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F46E;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F46E;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F46E;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F46E;&#x1F3FF;</span>
<span class="testElement">&#x1F46E;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F46E;&zwj;&#x2640;</span>
<span class="testElement">&#x1F46E;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F46E;&zwj;&#x2642;</span>
<span class="testElement">&#x1F46E;</span>
<span class="testElement">&#x1F471;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F471;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F471;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F471;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F471;&#x1F3FB;</span>
<span class="testElement">&#x1F471;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F471;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F471;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F471;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F471;&#x1F3FC;</span>
<span class="testElement">&#x1F471;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F471;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F471;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F471;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F471;&#x1F3FD;</span>
<span class="testElement">&#x1F471;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F471;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F471;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F471;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F471;&#x1F3FE;</span>
<span class="testElement">&#x1F471;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F471;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F471;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F471;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F471;&#x1F3FF;</span>
<span class="testElement">&#x1F471;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F471;&zwj;&#x2640;</span>
<span class="testElement">&#x1F471;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F471;&zwj;&#x2642;</span>
<span class="testElement">&#x1F471;</span>
<span class="testElement">&#x1F473;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F473;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F473;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F473;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F473;&#x1F3FB;</span>
<span class="testElement">&#x1F473;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F473;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F473;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F473;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F473;&#x1F3FC;</span>
<span class="testElement">&#x1F473;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F473;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F473;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F473;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F473;&#x1F3FD;</span>
<span class="testElement">&#x1F473;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F473;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F473;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F473;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F473;&#x1F3FE;</span>
<span class="testElement">&#x1F473;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F473;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F473;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F473;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F473;&#x1F3FF;</span>
<span class="testElement">&#x1F473;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F473;&zwj;&#x2640;</span>
<span class="testElement">&#x1F473;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F473;&zwj;&#x2642;</span>
<span class="testElement">&#x1F473;</span>
<span class="testElement">&#x1F477;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F477;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F477;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F477;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F477;&#x1F3FB;</span>
<span class="testElement">&#x1F477;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F477;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F477;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F477;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F477;&#x1F3FC;</span>
<span class="testElement">&#x1F477;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F477;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F477;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F477;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F477;&#x1F3FD;</span>
<span class="testElement">&#x1F477;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F477;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F477;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F477;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F477;&#x1F3FE;</span>
<span class="testElement">&#x1F477;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F477;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F477;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F477;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F477;&#x1F3FF;</span>
<span class="testElement">&#x1F477;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F477;&zwj;&#x2640;</span>
<span class="testElement">&#x1F477;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F477;&zwj;&#x2642;</span>
<span class="testElement">&#x1F477;</span>
<span class="testElement">&#x1F481;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F481;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F481;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F481;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F481;&#x1F3FB;</span>
<span class="testElement">&#x1F481;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F481;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F481;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F481;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F481;&#x1F3FC;</span>
<span class="testElement">&#x1F481;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F481;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F481;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F481;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F481;&#x1F3FD;</span>
<span class="testElement">&#x1F481;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F481;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F481;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F481;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F481;&#x1F3FE;</span>
<span class="testElement">&#x1F481;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F481;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F481;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F481;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F481;&#x1F3FF;</span>
<span class="testElement">&#x1F481;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F481;&zwj;&#x2640;</span>
<span class="testElement">&#x1F481;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F481;&zwj;&#x2642;</span>
<span class="testElement">&#x1F481;</span>
<span class="testElement">&#x1F482;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F482;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F482;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F482;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F482;&#x1F3FB;</span>
<span class="testElement">&#x1F482;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F482;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F482;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F482;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F482;&#x1F3FC;</span>
<span class="testElement">&#x1F482;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F482;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F482;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F482;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F482;&#x1F3FD;</span>
<span class="testElement">&#x1F482;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F482;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F482;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F482;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F482;&#x1F3FE;</span>
<span class="testElement">&#x1F482;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F482;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F482;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F482;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F482;&#x1F3FF;</span>
<span class="testElement">&#x1F482;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F482;&zwj;&#x2640;</span>
<span class="testElement">&#x1F482;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F482;&zwj;&#x2642;</span>
<span class="testElement">&#x1F482;</span>
<span class="testElement">&#x1F486;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F486;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F486;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F486;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F486;&#x1F3FB;</span>
<span class="testElement">&#x1F486;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F486;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F486;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F486;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F486;&#x1F3FC;</span>
<span class="testElement">&#x1F486;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F486;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F486;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F486;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F486;&#x1F3FD;</span>
<span class="testElement">&#x1F486;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F486;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F486;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F486;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F486;&#x1F3FE;</span>
<span class="testElement">&#x1F486;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F486;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F486;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F486;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F486;&#x1F3FF;</span>
<span class="testElement">&#x1F486;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F486;&zwj;&#x2640;</span>
<span class="testElement">&#x1F486;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F486;&zwj;&#x2642;</span>
<span class="testElement">&#x1F486;</span>
<span class="testElement">&#x1F487;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F487;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F487;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F487;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F487;&#x1F3FB;</span>
<span class="testElement">&#x1F487;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F487;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F487;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F487;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F487;&#x1F3FC;</span>
<span class="testElement">&#x1F487;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F487;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F487;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F487;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F487;&#x1F3FD;</span>
<span class="testElement">&#x1F487;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F487;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F487;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F487;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F487;&#x1F3FE;</span>
<span class="testElement">&#x1F487;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F487;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F487;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F487;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F487;&#x1F3FF;</span>
<span class="testElement">&#x1F487;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F487;&zwj;&#x2640;</span>
<span class="testElement">&#x1F487;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F487;&zwj;&#x2642;</span>
<span class="testElement">&#x1F487;</span>
<span class="testElement">&#x1F575;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F575;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F575;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F575;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F575;&#x1F3FB;</span>
<span class="testElement">&#x1F575;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F575;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F575;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F575;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F575;&#x1F3FC;</span>
<span class="testElement">&#x1F575;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F575;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F575;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F575;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F575;&#x1F3FD;</span>
<span class="testElement">&#x1F575;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F575;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F575;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F575;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F575;&#x1F3FE;</span>
<span class="testElement">&#x1F575;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F575;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F575;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F575;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F575;&#x1F3FF;</span>
<span class="testElement">&#x1F575;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F575;&zwj;&#x2640;</span>
<span class="testElement">&#x1F575;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F575;&zwj;&#x2642;</span>
<span class="testElement">&#x1F575;</span>
<span class="testElement">&#x1F645;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F645;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F645;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F645;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F645;&#x1F3FB;</span>
<span class="testElement">&#x1F645;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F645;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F645;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F645;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F645;&#x1F3FC;</span>
<span class="testElement">&#x1F645;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F645;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F645;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F645;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F645;&#x1F3FD;</span>
<span class="testElement">&#x1F645;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F645;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F645;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F645;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F645;&#x1F3FE;</span>
<span class="testElement">&#x1F645;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F645;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F645;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F645;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F645;&#x1F3FF;</span>
<span class="testElement">&#x1F645;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F645;&zwj;&#x2640;</span>
<span class="testElement">&#x1F645;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F645;&zwj;&#x2642;</span>
<span class="testElement">&#x1F645;</span>
<span class="testElement">&#x1F646;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F646;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F646;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F646;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F646;&#x1F3FB;</span>
<span class="testElement">&#x1F646;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F646;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F646;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F646;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F646;&#x1F3FC;</span>
<span class="testElement">&#x1F646;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F646;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F646;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F646;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F646;&#x1F3FD;</span>
<span class="testElement">&#x1F646;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F646;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F646;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F646;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F646;&#x1F3FE;</span>
<span class="testElement">&#x1F646;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F646;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F646;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F646;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F646;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F646;&zwj;&#x2640;</span>
<span class="testElement">&#x1F646;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F646;&zwj;&#x2642;</span>
<span class="testElement">&#x1F646;</span>
<span class="testElement">&#x1F647;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F647;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F647;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F647;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F647;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F647;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F647;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F647;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F647;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F647;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F647;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F647;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F647;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F647;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F647;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F647;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F647;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F647;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F647;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F647;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F647;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F647;&zwj;&#x2640;</span>
<span class="testElement">&#x1F647;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F647;&zwj;&#x2642;</span>
<span class="testElement">&#x1F64B;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F64B;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F64B;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F64B;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F64B;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F64B;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F64B;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F64B;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F64B;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F64B;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F64B;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F64B;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F64B;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F64B;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F64B;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F64B;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F64B;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F64B;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F64B;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F64B;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F64B;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F64B;&zwj;&#x2640;</span>
<span class="testElement">&#x1F64B;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F64B;&zwj;&#x2642;</span>
<span class="testElement">&#x1F64D;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F64D;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F64D;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F64D;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F64D;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F64D;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F64D;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F64D;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F64D;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F64D;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F64D;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F64D;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F64D;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F64D;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F64D;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F64D;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F64D;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F64D;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F64D;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F64D;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F64D;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F64D;&zwj;&#x2640;</span>
<span class="testElement">&#x1F64D;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F64D;&zwj;&#x2642;</span>
<span class="testElement">&#x1F64E;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F64E;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F64E;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F64E;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F64E;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F64E;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F64E;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F64E;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F64E;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F64E;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F64E;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F64E;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F64E;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F64E;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F64E;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F64E;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F64E;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F64E;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F64E;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F64E;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F64E;&#x1F3FF;</span>
<span class="testElement">&#x1F64E;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F64E;&zwj;&#x2640;</span>
<span class="testElement">&#x1F64E;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F64E;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6A3;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6A3;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6A3;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6A3;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6A3;&#x1F3FB;</span>
<span class="testElement">&#x1F6A3;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6A3;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6A3;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6A3;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6A3;&#x1F3FC;</span>
<span class="testElement">&#x1F6A3;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6A3;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6A3;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6A3;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6A3;&#x1F3FD;</span>
<span class="testElement">&#x1F6A3;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6A3;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6A3;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6A3;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6A3;&#x1F3FE;</span>
<span class="testElement">&#x1F6A3;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6A3;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6A3;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6A3;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6A3;&#x1F3FF;</span>
<span class="testElement">&#x1F6A3;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6A3;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6A3;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6A3;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6A3;</span>
<span class="testElement">&#x1F6B4;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6B4;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6B4;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6B4;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6B4;&#x1F3FB;</span>
<span class="testElement">&#x1F6B4;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6B4;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6B4;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6B4;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6B4;&#x1F3FC;</span>
<span class="testElement">&#x1F6B4;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6B4;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6B4;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6B4;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6B4;&#x1F3FD;</span>
<span class="testElement">&#x1F6B4;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6B4;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6B4;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6B4;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6B4;&#x1F3FE;</span>
<span class="testElement">&#x1F6B4;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6B4;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6B4;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6B4;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6B4;&#x1F3FF;</span>
<span class="testElement">&#x1F6B4;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6B4;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6B4;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6B4;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6B4;</span>
<span class="testElement">&#x1F6B5;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6B5;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6B5;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6B5;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6B5;&#x1F3FB;</span>
<span class="testElement">&#x1F6B5;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6B5;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6B5;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6B5;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6B5;&#x1F3FC;</span>
<span class="testElement">&#x1F6B5;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6B5;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6B5;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6B5;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6B5;&#x1F3FD;</span>
<span class="testElement">&#x1F6B5;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6B5;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6B5;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6B5;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6B5;&#x1F3FE;</span>
<span class="testElement">&#x1F6B5;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6B5;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6B5;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6B5;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6B5;&#x1F3FF;</span>
<span class="testElement">&#x1F6B5;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6B5;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6B5;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6B5;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6B5;</span>
<span class="testElement">&#x1F6B6;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6B6;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6B6;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6B6;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6B6;&#x1F3FB;</span>
<span class="testElement">&#x1F6B6;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6B6;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6B6;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6B6;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6B6;&#x1F3FC;</span>
<span class="testElement">&#x1F6B6;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6B6;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6B6;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6B6;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6B6;&#x1F3FD;</span>
<span class="testElement">&#x1F6B6;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6B6;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6B6;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6B6;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6B6;&#x1F3FE;</span>
<span class="testElement">&#x1F6B6;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6B6;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6B6;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6B6;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6B6;&#x1F3FF;</span>
<span class="testElement">&#x1F6B6;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6B6;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6B6;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6B6;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6B6;</span>
<span class="testElement">&#x26F9;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x26F9;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x26F9;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x26F9;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x26F9;&#x1F3FB;</span>
<span class="testElement">&#x26F9;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x26F9;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x26F9;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x26F9;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x26F9;&#x1F3FC;</span>
<span class="testElement">&#x26F9;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x26F9;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x26F9;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x26F9;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x26F9;&#x1F3FD;</span>
<span class="testElement">&#x26F9;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x26F9;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x26F9;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x26F9;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x26F9;&#x1F3FE;</span>
<span class="testElement">&#x26F9;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x26F9;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x26F9;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x26F9;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x26F9;&#x1F3FF;</span>
<span class="testElement">&#x26F9;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x26F9;&zwj;&#x2640;</span>
<span class="testElement">&#x26F9;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x26F9;&zwj;&#x2642;</span>
<span class="testElement">&#x26F9;</span>



<!--span class="testElement">&#x1F3C7;</span>
<span class="testElement">&#x1F3C7;&#x1F3FB;</span>
<span class="testElement">&#x1F3C7;&#x1F3FC;</span>
<span class="testElement">&#x1F3C7;&#x1F3FD;</span>
<span class="testElement">&#x1F3C7;&#x1F3FE;</span>
<span class="testElement">&#x1F3C7;&#x1F3FF;</span>
<span class="testElement">&#x1F3CC;</span>
<span class="testElement">&#x1F3CC;&#x1F3FB;</span>
<span class="testElement">&#x1F3CC;&#x1F3FC;</span>
<span class="testElement">&#x1F3CC;&#x1F3FD;</span>
<span class="testElement">&#x1F3CC;&#x1F3FE;</span>
<span class="testElement">&#x1F3CC;&#x1F3FF;</span>
<span class="testElement">&#x1F3CC;</span>
<span class="testElement">&#x1F574;&#x1F3FB;</span>
<span class="testElement">&#x1F574;&#x1F3FC;</span>
<span class="testElement">&#x1F574;&#x1F3FD;</span>
<span class="testElement">&#x1F574;&#x1F3FE;</span>
<span class="testElement">&#x1F574;&#x1F3FF;</span>
<span class="testElement">&#x1F3C2;</span>
<span class="testElement">&#x1F3C2;&#x1F3FB;</span>
<span class="testElement">&#x1F3C2;&#x1F3FC;</span>
<span class="testElement">&#x1F3C2;&#x1F3FD;</span>
<span class="testElement">&#x1F3C2;&#x1F3FE;</span>
<span class="testElement">&#x1F3C2;&#x1F3FF;</span>
<span class="testElement">&#x1F46A;</span>
<span class="testElement">&#x1F46A;&#x1F3FB;</span>
<span class="testElement">&#x1F46A;&#x1F3FC;</span>
<span class="testElement">&#x1F46A;&#x1F3FD;</span>
<span class="testElement">&#x1F46A;&#x1F3FE;</span>
<span class="testElement">&#x1F46A;&#x1F3FF;</span>
<span class="testElement">&#x1F46B;</span>
<span class="testElement">&#x1F46B;&#x1F3FB;</span>
<span class="testElement">&#x1F46B;&#x1F3FC;</span>
<span class="testElement">&#x1F46B;&#x1F3FD;</span>
<span class="testElement">&#x1F46B;&#x1F3FE;</span>
<span class="testElement">&#x1F46B;&#x1F3FF;</span>
<span class="testElement">&#x1F46C;</span>
<span class="testElement">&#x1F46C;&#x1F3FB;</span>
<span class="testElement">&#x1F46C;&#x1F3FC;</span>
<span class="testElement">&#x1F46C;&#x1F3FD;</span>
<span class="testElement">&#x1F46C;&#x1F3FE;</span>
<span class="testElement">&#x1F46C;&#x1F3FF;</span>
<span class="testElement">&#x1F46D;</span>
<span class="testElement">&#x1F46D;&#x1F3FB;</span>
<span class="testElement">&#x1F46D;&#x1F3FC;</span>
<span class="testElement">&#x1F46D;&#x1F3FD;</span>
<span class="testElement">&#x1F46D;&#x1F3FE;</span>
<span class="testElement">&#x1F46D;&#x1F3FF;</span>
<span class="testElement">&#x1F46F;</span>
<span class="testElement">&#x1F46F;&#x1F3FB;</span>
<span class="testElement">&#x1F46F;&#x1F3FC;</span>
<span class="testElement">&#x1F46F;&#x1F3FD;</span>
<span class="testElement">&#x1F46F;&#x1F3FE;</span>
<span class="testElement">&#x1F46F;&#x1F3FF;</span>
<span class="testElement">&#x1F6CC;</span>
<span class="testElement">&#x1F6CC;&#x1F3FB;</span>
<span class="testElement">&#x1F6CC;&#x1F3FC;</span>
<span class="testElement">&#x1F6CC;&#x1F3FD;</span>
<span class="testElement">&#x1F6CC;&#x1F3FE;</span>
<span class="testElement">&#x1F6CC;&#x1F3FF;</span-->
<span class="testElement">&#x1F926;</span>
<span class="testElement">&#x1F926;&#x1F3FB;</span>
<span class="testElement">&#x1F926;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F926;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F926;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F926;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F926;&zwj;&#x2640;</span>
<span class="testElement">&#x1F926;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F926;&zwj;&#x2642;</span>
<span class="testElement">&#x1F926;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F937;</span>
<span class="testElement">&#x1F937;&#x1F3FB;</span>
<span class="testElement">&#x1F937;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F937;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F937;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F937;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F937;&zwj;&#x2640;</span>
<span class="testElement">&#x1F937;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F937;&zwj;&#x2642;</span>
<span class="testElement">&#x1F937;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F938;</span>
<span class="testElement">&#x1F938;&#x1F3FB;</span>
<span class="testElement">&#x1F938;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F938;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F938;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F938;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F938;&zwj;&#x2640;</span>
<span class="testElement">&#x1F938;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F938;&zwj;&#x2642;</span>
<span class="testElement">&#x1F938;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F939;</span>
<span class="testElement">&#x1F939;&#x1F3FB;</span>
<span class="testElement">&#x1F939;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F939;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F939;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F939;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F939;&zwj;&#x2640;</span>
<span class="testElement">&#x1F939;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F939;&zwj;&#x2642;</span>
<span class="testElement">&#x1F939;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F93C;</span>
<span class="testElement">&#x1F93C;&#x1F3FB;</span>
<span class="testElement">&#x1F93C;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F93C;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F93C;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F93C;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F93C;&zwj;&#x2640;</span>
<span class="testElement">&#x1F93C;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F93C;&zwj;&#x2642;</span>
<span class="testElement">&#x1F93C;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F93D;</span>
<span class="testElement">&#x1F93D;&#x1F3FB;</span>
<span class="testElement">&#x1F93D;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F93D;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F93D;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F93D;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F93D;&zwj;&#x2640;</span>
<span class="testElement">&#x1F93D;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F93D;&zwj;&#x2642;</span>
<span class="testElement">&#x1F93D;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F93E;</span>
<span class="testElement">&#x1F93E;&#x1F3FB;</span>
<span class="testElement">&#x1F93E;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F93E;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F93E;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F93E;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F93E;&zwj;&#x2640;</span>
<span class="testElement">&#x1F93E;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F93E;&zwj;&#x2642;</span>
<span class="testElement">&#x1F93E;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F57A;</span>
<span class="testElement">&#x1F57A;&#x1F3FB;</span>
<span class="testElement">&#x1F919;</span>
<span class="testElement">&#x1F919;&#x1F3FB;</span>
<span class="testElement">&#x1F91A;</span>
<span class="testElement">&#x1F91A;&#x1F3FB;</span>
<span class="testElement">&#x1F91B;</span>
<span class="testElement">&#x1F91B;&#x1F3FB;</span>
<span class="testElement">&#x1F91C;</span>
<span class="testElement">&#x1F91C;&#x1F3FB;</span>
<span class="testElement">&#x1F91D;</span>
<span class="testElement">&#x1F91D;&#x1F3FB;</span>
<span class="testElement">&#x1F91E;</span>
<span class="testElement">&#x1F91E;&#x1F3FB;</span>
<span class="testElement">&#x1F930;</span>
<span class="testElement">&#x1F930;&#x1F3FB;</span>
<span class="testElement">&#x1F933;</span>
<span class="testElement">&#x1F933;&#x1F3FB;</span>
<span class="testElement">&#x1F934;</span>
<span class="testElement">&#x1F934;&#x1F3FB;</span>
<span class="testElement">&#x1F935;</span>
<span class="testElement">&#x1F935;&#x1F3FB;</span>
<span class="testElement">&#x1F936;</span>
<span class="testElement">&#x1F936;&#x1F3FB;</span>
<span class="testElement">&#x1F5A4;</span>
<span class="testElement">&#x1F6D1;</span>
<span class="testElement">&#x1F6D2;</span>
<span class="testElement">&#x1F6F4;</span>
<span class="testElement">&#x1F6F5;</span>
<span class="testElement">&#x1F6F6;</span>
<span class="testElement">&#x1F920;</span>
<span class="testElement">&#x1F921;</span>
<span class="testElement">&#x1F922;</span>
<span class="testElement">&#x1F923;</span>
<span class="testElement">&#x1F924;</span>
<span class="testElement">&#x1F925;</span>
<span class="testElement">&#x1F927;</span>
<span class="testElement">&#x1F93A;</span>
<span class="testElement">&#x1F940;</span>
<span class="testElement">&#x1F941;</span>
<span class="testElement">&#x1F942;</span>
<span class="testElement">&#x1F943;</span>
<span class="testElement">&#x1F944;</span>
<span class="testElement">&#x1F945;</span>
<span class="testElement">&#x1F947;</span>
<span class="testElement">&#x1F948;</span>
<span class="testElement">&#x1F949;</span>
<span class="testElement">&#x1F94A;</span>
<span class="testElement">&#x1F94B;</span>
<span class="testElement">&#x1F950;</span>
<span class="testElement">&#x1F951;</span>
<span class="testElement">&#x1F952;</span>
<span class="testElement">&#x1F953;</span>
<span class="testElement">&#x1F954;</span>
<span class="testElement">&#x1F955;</span>
<span class="testElement">&#x1F956;</span>
<span class="testElement">&#x1F957;</span>
<span class="testElement">&#x1F958;</span>
<span class="testElement">&#x1F959;</span>
<span class="testElement">&#x1F95A;</span>
<span class="testElement">&#x1F95B;</span>
<span class="testElement">&#x1F95C;</span>
<span class="testElement">&#x1F95D;</span>
<span class="testElement">&#x1F95E;</span>
<span class="testElement">&#x1F985;</span>
<span class="testElement">&#x1F986;</span>
<span class="testElement">&#x1F987;</span>
<span class="testElement">&#x1F988;</span>
<span class="testElement">&#x1F989;</span>
<span class="testElement">&#x1F98A;</span>
<span class="testElement">&#x1F98B;</span>
<span class="testElement">&#x1F98C;</span>
<span class="testElement">&#x1F98D;</span>
<span class="testElement">&#x1F98E;</span>
<span class="testElement">&#x1F98F;</span>
<span class="testElement">&#x1F990;</span>
<span class="testElement">&#x1F991;</span>
<span class="testElement">&#x1F468;</span>
<span class="testElement">&#x1F468;&zwj;&#x2695;&#xFE0F;</span>
<span class="testElement">&#x1F468;&#x1F3FB;&zwj;&#x2695;&#xFE0F;</span>
<span class="testElement">&#x1F468;&zwj;&#x2695;</span>
<span class="testElement">&#x1F468;&#x1F3FB;&zwj;&#x2695;</span>
<span class="testElement">&#x1F469;</span>
<span class="testElement">&#x1F469;&zwj;&#x2695;&#xFE0F;</span>
<span class="testElement">&#x1F469;&#x1F3FB;&zwj;&#x2695;&#xFE0F;</span>
<span class="testElement">&#x1F469;&zwj;&#x2695;</span>
<span class="testElement">&#x1F469;&#x1F3FB;&zwj;&#x2695;</span>
<span class="testElement">&#x1F468;&zwj;&#x1F33E;&#xFE0F;</span>
<span class="testElement">&#x1F468;&#x1F3FB;&zwj;&#x1F33E;&#xFE0F;</span>
<span class="testElement">&#x1F468;&zwj;&#x1F33E;</span>
<span class="testElement">&#x1F468;&#x1F3FB;&zwj;&#x1F33E;</span>
<span class="testElement">&#x1F469;&zwj;&#x1F33E;&#xFE0F;</span>
<span class="testElement">&#x1F469;&#x1F3FB;&zwj;&#x1F33E;&#xFE0F;</span>
<span class="testElement">&#x1F469;&zwj;&#x1F33E;</span>
<span class="testElement">&#x1F469;&#x1F3FB;&zwj;&#x1F33E;</span>
<span class="testElement">&#x1F468;&zwj;&#x1F373;&#xFE0F;</span>
<span class="testElement">&#x1F468;&#x1F3FB;&zwj;&#x1F373;&#xFE0F;</span>
<span class="testElement">&#x1F468;&zwj;&#x1F373;</span>
<span class="testElement">&#x1F468;&#x1F3FB;&zwj;&#x1F373;</span>
<span class="testElement">&#x1F469;&zwj;&#x1F373;&#xFE0F;</span>
<span class="testElement">&#x1F469;&#x1F3FB;&zwj;&#x1F373;&#xFE0F;</span>
<span class="testElement">&#x1F469;&zwj;&#x1F373;</span>
<span class="testElement">&#x1F469;&#x1F3FB;&zwj;&#x1F373;</span>
<span class="testElement">&#x1F468;&zwj;&#x1F393;&#xFE0F;</span>
<span class="testElement">&#x1F468;&#x1F3FB;&zwj;&#x1F393;&#xFE0F;</span>
<span class="testElement">&#x1F468;&zwj;&#x1F393;</span>
<span class="testElement">&#x1F468;&#x1F3FB;&zwj;&#x1F393;</span>
<span class="testElement">&#x1F469;&zwj;&#x1F393;&#xFE0F;</span>
<span class="testElement">&#x1F469;&#x1F3FB;&zwj;&#x1F393;&#xFE0F;</span>
<span class="testElement">&#x1F469;&zwj;&#x1F393;</span>
<span class="testElement">&#x1F469;&#x1F3FB;&zwj;&#x1F393;</span>
<span class="testElement">&#x1F468;&zwj;&#x1F3A4;&#xFE0F;</span>
<span class="testElement">&#x1F468;&#x1F3FB;&zwj;&#x1F3A4;&#xFE0F;</span>
<span class="testElement">&#x1F468;&zwj;&#x1F3A4;</span>
<span class="testElement">&#x1F468;&#x1F3FB;&zwj;&#x1F3A4;</span>
<span class="testElement">&#x1F469;&zwj;&#x1F3A4;&#xFE0F;</span>
<span class="testElement">&#x1F469;&#x1F3FB;&zwj;&#x1F3A4;&#xFE0F;</span>
<span class="testElement">&#x1F469;&zwj;&#x1F3A4;</span>
<span class="testElement">&#x1F469;&#x1F3FB;&zwj;&#x1F3A4;</span>
<span class="testElement">&#x1F468;&zwj;&#x1F3EB;&#xFE0F;</span>
<span class="testElement">&#x1F468;&#x1F3FB;&zwj;&#x1F3EB;&#xFE0F;</span>
<span class="testElement">&#x1F468;&zwj;&#x1F3EB;</span>
<span class="testElement">&#x1F468;&#x1F3FB;&zwj;&#x1F3EB;</span>
<span class="testElement">&#x1F469;&zwj;&#x1F3EB;&#xFE0F;</span>
<span class="testElement">&#x1F469;&#x1F3FB;&zwj;&#x1F3EB;&#xFE0F;</span>
<span class="testElement">&#x1F469;&zwj;&#x1F3EB;</span>
<span class="testElement">&#x1F469;&#x1F3FB;&zwj;&#x1F3EB;</span>
<span class="testElement">&#x1F468;&zwj;&#x1F3ED;&#xFE0F;</span>
<span class="testElement">&#x1F468;&#x1F3FB;&zwj;&#x1F3ED;&#xFE0F;</span>
<span class="testElement">&#x1F468;&zwj;&#x1F3ED;</span>
<span class="testElement">&#x1F468;&#x1F3FB;&zwj;&#x1F3ED;</span>
<span class="testElement">&#x1F469;&zwj;&#x1F3ED;&#xFE0F;</span>
<span class="testElement">&#x1F469;&#x1F3FB;&zwj;&#x1F3ED;&#xFE0F;</span>
<span class="testElement">&#x1F469;&zwj;&#x1F3ED;</span>
<span class="testElement">&#x1F469;&#x1F3FB;&zwj;&#x1F3ED;</span>
<span class="testElement">&#x1F468;&zwj;&#x1F4BB;&#xFE0F;</span>
<span class="testElement">&#x1F468;&#x1F3FB;&zwj;&#x1F4BB;&#xFE0F;</span>
<span class="testElement">&#x1F468;&zwj;&#x1F4BB;</span>
<span class="testElement">&#x1F468;&#x1F3FB;&zwj;&#x1F4BB;</span>
<span class="testElement">&#x1F469;&zwj;&#x1F4BB;&#xFE0F;</span>
<span class="testElement">&#x1F469;&#x1F3FB;&zwj;&#x1F4BB;&#xFE0F;</span>
<span class="testElement">&#x1F469;&zwj;&#x1F4BB;</span>
<span class="testElement">&#x1F469;&#x1F3FB;&zwj;&#x1F4BB;</span>
<span class="testElement">&#x1F468;&zwj;&#x1F4BC;&#xFE0F;</span>
<span class="testElement">&#x1F468;&#x1F3FB;&zwj;&#x1F4BC;&#xFE0F;</span>
<span class="testElement">&#x1F468;&zwj;&#x1F4BC;</span>
<span class="testElement">&#x1F468;&#x1F3FB;&zwj;&#x1F4BC;</span>
<span class="testElement">&#x1F469;&zwj;&#x1F4BC;&#xFE0F;</span>
<span class="testElement">&#x1F469;&#x1F3FB;&zwj;&#x1F4BC;&#xFE0F;</span>
<span class="testElement">&#x1F469;&zwj;&#x1F4BC;</span>
<span class="testElement">&#x1F469;&#x1F3FB;&zwj;&#x1F4BC;</span>
<span class="testElement">&#x1F468;&zwj;&#x1F527;&#xFE0F;</span>
<span class="testElement">&#x1F468;&#x1F3FB;&zwj;&#x1F527;&#xFE0F;</span>
<span class="testElement">&#x1F468;&zwj;&#x1F527;</span>
<span class="testElement">&#x1F468;&#x1F3FB;&zwj;&#x1F527;</span>
<span class="testElement">&#x1F469;&zwj;&#x1F527;&#xFE0F;</span>
<span class="testElement">&#x1F469;&#x1F3FB;&zwj;&#x1F527;&#xFE0F;</span>
<span class="testElement">&#x1F469;&zwj;&#x1F527;</span>
<span class="testElement">&#x1F469;&#x1F3FB;&zwj;&#x1F527;</span>
<span class="testElement">&#x1F468;&zwj;&#x1F52C;&#xFE0F;</span>
<span class="testElement">&#x1F468;&#x1F3FB;&zwj;&#x1F52C;&#xFE0F;</span>
<span class="testElement">&#x1F468;&zwj;&#x1F52C;</span>
<span class="testElement">&#x1F468;&#x1F3FB;&zwj;&#x1F52C;</span>
<span class="testElement">&#x1F469;&zwj;&#x1F52C;&#xFE0F;</span>
<span class="testElement">&#x1F469;&#x1F3FB;&zwj;&#x1F52C;&#xFE0F;</span>
<span class="testElement">&#x1F469;&zwj;&#x1F52C;</span>
<span class="testElement">&#x1F469;&#x1F3FB;&zwj;&#x1F52C;</span>
<span class="testElement">&#x1F468;&zwj;&#x1F3A8;&#xFE0F;</span>
<span class="testElement">&#x1F468;&#x1F3FB;&zwj;&#x1F3A8;&#xFE0F;</span>
<span class="testElement">&#x1F468;&zwj;&#x1F3A8;</span>
<span class="testElement">&#x1F468;&#x1F3FB;&zwj;&#x1F3A8;</span>
<span class="testElement">&#x1F469;&zwj;&#x1F3A8;&#xFE0F;</span>
<span class="testElement">&#x1F469;&#x1F3FB;&zwj;&#x1F3A8;&#xFE0F;</span>
<span class="testElement">&#x1F469;&zwj;&#x1F3A8;</span>
<span class="testElement">&#x1F469;&#x1F3FB;&zwj;&#x1F3A8;</span>
<span class="testElement">&#x1F468;&zwj;&#x1F692;&#xFE0F;</span>
<span class="testElement">&#x1F468;&#x1F3FB;&zwj;&#x1F692;&#xFE0F;</span>
<span class="testElement">&#x1F468;&zwj;&#x1F692;</span>
<span class="testElement">&#x1F468;&#x1F3FB;&zwj;&#x1F692;</span>
<span class="testElement">&#x1F469;&zwj;&#x1F692;&#xFE0F;</span>
<span class="testElement">&#x1F469;&#x1F3FB;&zwj;&#x1F692;&#xFE0F;</span>
<span class="testElement">&#x1F469;&zwj;&#x1F692;</span>
<span class="testElement">&#x1F469;&#x1F3FB;&zwj;&#x1F692;</span>
<span class="testElement">&#x1F468;&zwj;&#x2708;&#xFE0F;</span>
<span class="testElement">&#x1F468;&#x1F3FB;&zwj;&#x2708;&#xFE0F;</span>
<span class="testElement">&#x1F468;&zwj;&#x2708;</span>
<span class="testElement">&#x1F468;&#x1F3FB;&zwj;&#x2708;</span>
<span class="testElement">&#x1F469;&zwj;&#x2708;&#xFE0F;</span>
<span class="testElement">&#x1F469;&#x1F3FB;&zwj;&#x2708;&#xFE0F;</span>
<span class="testElement">&#x1F469;&zwj;&#x2708;</span>
<span class="testElement">&#x1F469;&#x1F3FB;&zwj;&#x2708;</span>
<span class="testElement">&#x1F468;&zwj;&#x1F680;&#xFE0F;</span>
<span class="testElement">&#x1F468;&#x1F3FB;&zwj;&#x1F680;&#xFE0F;</span>
<span class="testElement">&#x1F468;&zwj;&#x1F680;</span>
<span class="testElement">&#x1F468;&#x1F3FB;&zwj;&#x1F680;</span>
<span class="testElement">&#x1F469;&zwj;&#x1F680;&#xFE0F;</span>
<span class="testElement">&#x1F469;&#x1F3FB;&zwj;&#x1F680;&#xFE0F;</span>
<span class="testElement">&#x1F469;&zwj;&#x1F680;</span>
<span class="testElement">&#x1F469;&#x1F3FB;&zwj;&#x1F680;</span>
<span class="testElement">&#x1F468;&zwj;&#x2708;&#xFE0F;</span>
<span class="testElement">&#x1F468;&#x1F3FB;&zwj;&#x2708;&#xFE0F;</span>
<span class="testElement">&#x1F468;&zwj;&#x2708;</span>
<span class="testElement">&#x1F468;&#x1F3FB;&zwj;&#x2708;</span>
<span class="testElement">&#x1F469;&zwj;&#x2708;&#xFE0F;</span>
<span class="testElement">&#x1F469;&#x1F3FB;&zwj;&#x2708;&#xFE0F;</span>
<span class="testElement">&#x1F469;&zwj;&#x2708;</span>
<span class="testElement">&#x1F469;&#x1F3FB;&zwj;&#x2708;</span>
<span class="testElement">&#x1F93C;</span>
<span class="testElement">&#x1F93C;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F93C;&zwj;&#x2640;</span>
<span class="testElement">&#x1F93C;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F93C;&zwj;&#x2642;</span>
<script>
var controlWidth = document.getElementById("control").offsetWidth;
let elements = document.getElementsByClassName("testElement");
for (var i = 0; i < elements.length; ++i) {
	shouldBe("elements[i].offsetWidth", "controlWidth");
}
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
